+++ frontend/src/metabase/visualizations/visualizations/TableEditable/index.ts
@@ -1,0 +1,1 @@
+export { TableEditable } from "./TableEditable";
